TVGUI-13 Context wrapper
