+++ src/contexts/index.ts
@@ -2,10 +2,12 @@ import { RecordedShowsContext } from "./RecordedShowsContext";
 import { RemindersContext } from "./RemindersContext";
 import { SearchListContext } from "./SearchListContext";
 import { UserContext } from "./UserContext";
+import ContextWrapper from "./ContextWrapper";
 
 export {
     RecordedShowsContext,
     RemindersContext,
     SearchListContext,
-    UserContext
+    UserContext,
+    ContextWrapper
 };
